Fix: 상대페이지에서 팔로우 버튼 활성화
