+++ src/pages/YourProfile.jsx
@@ -1,4 +1,4 @@
-import { React, useState } from 'react';
+import { React, useState, useEffect } from 'react';
 import { useLocation } from "react-router-dom";
 import UserProfile from '../components/profile/userProfile'
 import UserProduct from '../components/profile/userProduct';
@@ -8,7 +8,7 @@ import TabMenu from '../components/common/tabMenu';
 import { AlertLogoutModal } from '../components/common/alert'
 import { ProfileSection } from '../components/profile/userProfile/index.style';
 
-function YourProfilePage() {
+function YourProfilePage(props) {
     const [alertOn, setAlertOn] = useState(false);
     function alertOnModal() {
         setAlertOn(true);
@@ -27,11 +27,41 @@ function YourProfilePage() {
 
     const accountname = checkAccountName();
 
+    //data 받아오기
+    const [userData, setUserData] = useState();
+    const token = sessionStorage.getItem('token');
+
+    const getData = async (account) => {
+        const res = await fetch(`https://mandarin.api.weniv.co.kr/profile/${account}`, {
+            method: "GET",
+            headers: {
+                "Authorization": `Bearer ${token}`,
+                "Content-type": "application/json"
+            }
+        })
+        const json = await res.json()
+        setUserData(json)
+        console.log('데이터', userData)
+    }
+
+    useEffect(() => {
+        if (props.accountname) {
+            getData(props.accountname)
+        }
+        else {
+            getData(accountname)
+        }
+    }, [])
+
+    if (!userData) {
+        return <></>
+    }
+
     return (
         <>
             <TopBasicNav alertOnModal={alertOnModal} />
             <ProfileSection>
-                <UserProfile accountname={accountname} />
+                <UserProfile userData={userData} />
                 <UserProduct accountname={accountname} />
                 <UserPost accountname={accountname} />
             </ProfileSection>
